Fixing windows tests.
Here we need to have an exe to run on windows, since golang is easier to
build on multiple platforms than rust, I chose it. I'm checking in the
exe's since we need them to be able to run without having to install
golang everywhere. We could put them up into artifactory so they aren't
checked in but it's only 1.6M each. This is something we can discuss
before merging.

Here I also took the OperatingSystem mostly intact from Gradle. The one
change that I did was to remove the extension stipping, since everything
shouldn't have one if possible.

I split the PythonDetailsTest into PythonDetailsUnixTest and
PythonDetailsWindowsTest to make sure that everything can be tested on
different platforms.